Crypto Education for 3rd World
Zolu Project aims to spread crypto education to enhance the awareness about crypto.
Specially in 3rd world countries, where it’s still a myth. as we mentioned in Zolu white paper.
https://zoluc.com/
We will start approaching the local community with local language influencers. In this regard we will marketing the zolu coin to cover the expenses of project and beside in all social media influencer promotions they explain them about crypto